# Optimisation: only traverse AST's if the cache is stale. 