--- a/sway-lsp/benches/lsp_benchmarks/compile.rs
+++ b/sway-lsp/benches/lsp_benchmarks/compile.rs
@@ -44,7 +44,8 @@ fn benchmarks(c: &mut Criterion) {
         );
         let session = Arc::new(session::Session::new());
         let member_path = sync.member_path(&uri).unwrap();
-
+        let modified_file = sway_lsp::server_state::modified_file(lsp_mode.as_ref());
+        
         b.iter(|| {
             let _ = black_box(
                 session::traverse(
@@ -54,7 +55,7 @@ fn benchmarks(c: &mut Criterion) {
                     &engines_clone,
                     session.clone(),
                     &state.token_map,
-                    lsp_mode.as_ref(),
+                    modified_file.as_ref(),
                 )
                 .unwrap(),
             );