## Description
This PR introduces selective processing to dramatically improve LSP performance when opening multiple files.

### Key Changes
* **Conditional reprocessing**: Only perform expensive traversal operations when tokens don't exist or files have been modified
* **Selective garbage collection**: GC now only runs when files are actually modified, not on every compilation
* **Targeted token map updates**: Remove tokens only for modified files instead of clearing the entire map
* **Improved modified file tracking**: Extract modified file detection into reusable utility functions

### Performance Impact
| Test | Before | After | Improvement |
|------|--------|-------|-------------|
| did_open_all_std_lib_files | 54.96s | 3.76s | 14.6x faster |
| Time reduction | - | - | 93% decrease |

The `did_open_all_std_lib_files` test opens every file in the std lib and publishes the semantic tokens for the file. 

The optimisation is particularly effective for workflows involving opening multiple files where the LSP previously performed unnecessary recompilation and traversal for unchanged files.

## Checklist

- [ ] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand areas.
- [ ] I have updated the documentation where relevant (API docs, the reference, and the Sway book).
   - [ ] If my change requires substantial documentation changes, I have [requested support from the DevRel team](https://github.com/FuelLabs/devrel-requests/issues/new/choose)
- [x] I have added tests that prove my fix is effective or that my feature works.
- [x] I have added (or requested a maintainer to add) the necessary `Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs Code Review Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.